requirements.txt
# requirements.txt
Flask
numpy
Pillow
tensorflow
jinja2
google-auth-oauthlib
google-api-python-client
requests